libdm: remove visualizer
Removed from main branch and moved to visualizer branch. For the
visualizer to work correctly is has to be a separate process, otherwise
it will also display its own memory e.g. SDL2.

After the main "alloc" functions are re-written I plan to re-write the
visualizer, but for now it has been moved to separate branch.

Signed-off-by: Epicurius <niklas0702@gmail.com>