free: improve 'merge_free_adjacent_blocks()'
Make sure to get the real block size when defragmenting. Because, there might
be unused memory between blocks. This happens when reserving a free block which
is between two reserved blocks, and the block size is larger than the required
size, but to small for splitting into two separate blocks.

Signed-off-by: Epicurius <niklas0702@gmail.com>