tests: add benchmark script and test
By running "./tests/run_benchmark.py" a benchmark test will be run
using 'libdm' and 'stdlib' and the results will be printed to the
standard output. Additionally when '--commit' options is passed
the results will be written to './tests/benchmarklog' and git commited.
This will help keep track of performance changes.

Signed-off-by: Epicurius <niklas.neronin@gmail.com>